Fix: Cast float to int for ImageResizer::image_resize() parameter (#262)
* docs: 幅・高さは intを想定しています

* docs: 戻り値は true|false なので bool型です

* Fix: Cast float to int for ImageResizer::image_resize() parameter

- Convert $width * 1.5 to (int)round($width * 1.5) in proc_image_resize()

- Resolves PHPStan error: expects integer but float was provided for $i_maxH parameter

+++ smartybook/chapter5_3/plib/imgsizecvt_funcs.php
@@ -84,7 +84,7 @@ function proc_image_resize($i_fname)
     if (file_exists($src_path)) {
         foreach ([120, 240, 480] as $width) {
             $dst_path = $CFG['DSTIMG_DIR'] . "$width/$fname";
-            ImageResizer::image_resize($src_path, $dst_path, $width, $width * 1.5);
+            ImageResizer::image_resize($src_path, $dst_path, $width, (int)($width * 1.5));
         }
     }
 }
